+++ .dart_tool/package_config_subset
@@ -108,8 +108,8 @@ file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.12/
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.12/lib/
 material_color_utilities
 2.13
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.1.5/
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.1.5/lib/
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.2.0/
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.2.0/lib/
 meta
 2.12
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/meta-1.8.0/
@@ -136,16 +136,16 @@ file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/plugin_platform_interf
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/plugin_platform_interface-2.1.3/lib/
 source_span
 2.14
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/source_span-1.9.0/
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/source_span-1.9.0/lib/
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/source_span-1.9.1/
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/source_span-1.9.1/lib/
 stack_trace
 2.12
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/lib/
 stream_channel
-2.12
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/lib/
+2.14
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.1/
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.1/lib/
 string_scanner
 2.12
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.1/
@@ -155,17 +155,17 @@ term_glyph
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.1/
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.1/lib/
 test_api
-2.12
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.12/
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.12/lib/
+2.18
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.14/
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.14/lib/
 typed_data
 2.12
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/typed_data-1.3.1/
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/typed_data-1.3.1/lib/
 vector_math
 2.14
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.2/
-file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.2/lib/
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.4/
+file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.4/lib/
 xml
 2.17
 file:///home/codespace/.pub-cache/hosted/pub.dartlang.org/xml-6.1.0/
@@ -187,7 +187,7 @@ flutter_web_plugins
 file:///home/codespace/flutter/packages/flutter_web_plugins/
 file:///home/codespace/flutter/packages/flutter_web_plugins/lib/
 helloworld
-2.12
+2.17
 file:///workspaces/APP_FreshPlus_TakeCareMyRefrigerator/
 file:///workspaces/APP_FreshPlus_TakeCareMyRefrigerator/lib/
 2
