+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"cloud_firestore","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-3.4.8/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_auth","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_auth-3.9.0/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_core-1.22.0/","native_build":true,"dependencies":[]}],"android":[{"name":"cloud_firestore","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-3.4.8/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_auth","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_auth-3.9.0/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_core-1.22.0/","native_build":true,"dependencies":[]}],"macos":[{"name":"cloud_firestore","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-3.4.8/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_auth","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_auth-3.9.0/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_core-1.22.0/","native_build":true,"dependencies":[]}],"linux":[],"windows":[],"web":[{"name":"cloud_firestore_web","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/cloud_firestore_web-2.8.7/","dependencies":["firebase_core_web"]},{"name":"firebase_auth_web","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_auth_web-4.4.1/","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_core_web-1.7.2/","dependencies":[]}]},"dependencyGraph":[{"name":"cloud_firestore","dependencies":["cloud_firestore_web","firebase_core"]},{"name":"cloud_firestore_web","dependencies":["firebase_core","firebase_core_web"]},{"name":"firebase_auth","dependencies":["firebase_auth_web","firebase_core"]},{"name":"firebase_auth_web","dependencies":["firebase_core","firebase_core_web"]},{"name":"firebase_core","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","dependencies":[]}],"date_created":"2022-09-25 10:03:00.683030","version":"3.3.2"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"cloud_firestore","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-3.4.8/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_auth","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_auth-3.9.0/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_core-1.22.0/","native_build":true,"dependencies":[]}],"android":[{"name":"cloud_firestore","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-3.4.8/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_auth","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_auth-3.9.0/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_core-1.22.0/","native_build":true,"dependencies":[]}],"macos":[{"name":"cloud_firestore","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-3.4.8/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_auth","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_auth-3.9.0/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_core-1.22.0/","native_build":true,"dependencies":[]}],"linux":[],"windows":[],"web":[{"name":"cloud_firestore_web","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/cloud_firestore_web-2.8.7/","dependencies":["firebase_core_web"]},{"name":"firebase_auth_web","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_auth_web-4.4.1/","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","path":"/home/codespace/.pub-cache/hosted/pub.dartlang.org/firebase_core_web-1.7.2/","dependencies":[]}]},"dependencyGraph":[{"name":"cloud_firestore","dependencies":["cloud_firestore_web","firebase_core"]},{"name":"cloud_firestore_web","dependencies":["firebase_core","firebase_core_web"]},{"name":"firebase_auth","dependencies":["firebase_auth_web","firebase_core"]},{"name":"firebase_auth_web","dependencies":["firebase_core","firebase_core_web"]},{"name":"firebase_core","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","dependencies":[]}],"date_created":"2022-09-30 03:08:27.890887","version":"3.4.0-34.0.pre.31"}
