+++ .vscode/extensions.json
@@ -1,0 +1,5 @@
+{
+    "recommendations": [
+        "Dart-Code.flutter"
+    ]
+}
