+++ pubspec.yaml
@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   cloud_firestore: ^3.4.8
@@ -22,6 +22,7 @@ dependencies:
   firebase_core: ^1.22.0
   flutter:
     sdk: flutter
+  get: ^4.6.5
 
 dev_dependencies:
   flutter_test:
@@ -33,7 +34,6 @@ dev_dependencies:
   bottom_navy_bar: ^6.0.0
   firebase_auth: ^3.9.0
   logger: ^1.1.0
-  get: ^4.6.5
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
